refactor: AttachmentFile 통합에 따른 디렉토리 수정
- 사용하지 않는 파일 삭제
