+++ src/main/java/com/careerfit/attachmentfile/dto/PresignedUrlResponse.java
@@ -1,4 +1,4 @@
-package com.careerfit.document.dto;
+package com.careerfit.attachmentfile.dto;
 
 public record PresignedUrlResponse(String presignedUrl, String uniqueFileName) {
 }
