+++ src/main/java/com/careerfit/global/config/AwsProperties.java
@@ -1,4 +1,4 @@
-package com.careerfit.document.config;
+package com.careerfit.global.config;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
