--- src/main/java/com/careerfit/resume/repository/ResumeRepository.java
@@ -1,8 +1,0 @@
-package com.careerfit.resume.repository;
-
-import com.careerfit.resume.domain.Resume;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ResumeRepository extends JpaRepository<Resume, Long> {
-
-}
